requirements.txt
Django==2.2.6
djangorestframework==3.10.3
psycopg2-binary==2.8.3
